module_2/requirements.txt
#logging==0.5.1.2
PyYAML==6.0
vk_api==11.9.9
unittest2==1.1.0
